refactor(runtime): split runtime live state
